+++ lib/cn/meson.build
@@ -27,7 +27,6 @@ cn_sources = files(
     'vcomp_params.c',
     'wbt_builder.c',
     'wbt_reader.c',
-    'wbt_reader_v6.c',
 )
 
 cn_includes = include_directories('.')
